requirements.txt
Pillow
numpy
pytesseract
rapidfuzz
pyperclip
pynput
scipy
pystray
pyqt6
requests